clean up timeline card files
